% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_tc.R
\name{cl_tc}
\alias{cl_tc}
\title{Component Loadings for Temporal Consistency}
\usage{
cl_tc(it1, it2)
}
\arguments{
\item{it1}{Columns in a dataframe or matrix containing the items measured at time 1}

\item{it2}{Columns in a dataframe or matrix containing the items measured at time 2}
}
\value{
Temporal consistency estimated as the similarity across component loadings
}
\description{
This function estimates item-level temporal consistency
using component loadings as recommended by DeSimone(2015).
The input consistes of two data frames containing the focal items
of interest at time 1 (it1) and time 2 (it2).
}
