% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tc.R
\name{cor_tc}
\alias{cor_tc}
\title{Test-Level and Item-Level Temporal Consistency Correlations}
\usage{
cor_tc(tot1, tot2, it1, it2)
}
\arguments{
\item{tot1}{Total test score at time 1}

\item{tot2}{Total test score at time 1}

\item{it1}{Item scores at time 1}

\item{it2}{Item scores at time 2}
}
\value{
Estimates of item-level and test-level test-retest correlations
}
\description{
This function estimates temporal consistency using p x i x o design in G-Theory
as recommended by DeSimone(2015). The input consistes of two data frames
containing the focal items of interest at time 1 (df1) and time 2 (df2).
}
