% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ptc.R
\name{d_ptc}
\alias{d_ptc}
\title{D Estimate for Personal Temporal Consistency}
\usage{
d_ptc(id, it1, it2)
}
\arguments{
\item{id}{Columne in a dataframe or matrix containing the subject ID variable}

\item{it1}{Columns in a dataframe or matrix containing the items measured at time 1}

\item{it2}{Columns in a dataframe or matrix containing the items measured at time 2}
}
\value{
D and D2 values for personal temporal consistency
}
\description{
This function estimates person-level temporal consistency
  by estimating Mahalanobis D and D^2 statistics for personal
  temporal consistency as recommended by DeSimone(2015).
  The D^2 statistic is distributed as a Chi-Square with
  degrees of freedom equal to the number of focal items.
  The input consistes of the focal items measured at
  time 1 (it1) and time 2 (it2).
}
\details{
Link to DeSimone(2015): https://journals.sagepub.com/doi/full/10.1177/1094428114553061
}
\examples{
data(desimone1)
names(desimone1)
d_ptc(desimone1[,1], desimone1[,c(2:6)],desimone1[,c(8:12)])

}
