% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desimone2.R
\docType{data}
\name{desimone2}
\alias{desimone2}
\title{Data from DeSimone (2015)'s Example 2}
\format{
A data frame with 30 rows and 12 variables:
\describe{
  \item{id}{subject id number}
  \item{i11}{responses to item 1 measured at time 1}
  \item{i21}{responses to item 2 measured at time 1}
  \item{i31}{responses to item 3 measured at time 1}
  \item{i41}{responses to item 4 measured at time 1}
  \item{i51}{responses to item 5 measured at time 1}
  \item{t1}{total score at time 1 (i.e., unit-weighted composite of i11:51)}
  \item{i12}{responses to item 1 measured at time 2}
  \item{i22}{responses to item 2 measured at time 2...}
  \item{i32}{responses to item 3 measured at time 2...}
  \item{i42}{responses to item 4 measured at time 2...}
  \item{i52}{responses to item 5 measured at time 2...}
  \item{t2}{total score at time 2 (i.e., unit-weighted composite of i12:52)}
}
}
\usage{
desimone2
}
\description{
A dataset containing the simulated responses of 30 individuals to
5 items measured two times.
}
\keyword{datasets}
