% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmr_tc.R
\name{srmr_tc}
\alias{srmr_tc}
\title{Square Root Mean Residual for Temporal Consistency}
\usage{
srmr_tc(it1, it2)
}
\arguments{
\item{it1}{Columns in a dataframe or matrix containing the items measured at time 1}

\item{it2}{Columns in a dataframe or matrix containing the items measured at time 2}
}
\value{
Estimate of SRMR for item-level test-retest correlations
}
\description{
This function estimates item-level temporal consistency using
the square root mean residual (SRMR) for as recommended by DeSimone(2015).
The input consistes of two data frames containing the focal items
of interest at time 1 (df1) and time 2 (df2).
}
