% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_all.R
\name{tc_all}
\alias{tc_all}
\title{Estimates of Temporal Consistency from DeSimone (2015)}
\usage{
tc_all(id, it1, tot1, it2, tot2, printD2 = F)
}
\arguments{
\item{id}{Case ID variable}

\item{it1}{Item scores at time 1}

\item{tot1}{Total test score at time 1}

\item{it2}{Item scores at time 2}

\item{tot2}{Total test score at time 1}
}
\value{
List of results
}
\description{
This function estimates test-level and item-level temporal consistency
using: test-retest correlations, component loadings, square root mean
residual, and D^2 for personal temporal consistency.
}
